+++ src/js/base.js
@@ -12,28 +12,17 @@ window.addEventListener("load", () => {
 
         window.addEventListener("beforeunload", (e) => {
             const settingsFile = path.join(app.mainFolder, "data", "settings.json");
-            fsp.readFile(settingsFile, "utf-8").then((data) => {
+            fsp.readFile(settingsFile, "utf8").then((data) => {
                 const jsonData = JSON.parse(data);
 
-                const saveData = {
-                    volume: app.elements.footer.volume.slider.value / 100,
-                    loop: app.settings.loop,
-                    random: app.settings.random,
-                    window: {
-                        x: Math.max(app.window.x, 0),
-                        y: Math.max(app.window.y, 0),
-                        w: app.window.w,
-                        h: app.window.h,
-                        f: app.window.f,
-                    },
-                    playlists: [],
-                    colors: jsonData.colors,
-                };
-                const strSaveData = JSON.stringify(saveData, null, 2);
+                if (app.settings.window.x < 0) app.settings.window.x = 0; 
+                if (app.settings.window.y < 0) app.settings.window.y = 0; 
+
+                const strSaveData = JSON.stringify(app.settings, null, 2);
                 fs.writeFile(settingsFile, strSaveData, (err) => {
-                    if (err) console.error("Error writing json:", err);
+                    if (err) console.error("ERROR HK-217 => Could not write settings.json:", err);
                 });
-            }).catch((readErr) => console.error("Error => can't read songs.json:" + readErr));
+            }).catch((readErr) => console.error("ERROR HK-116 => Could not read settings.json:" + readErr));
         });
     });
 });