+++ src/js/song-listener.js
@@ -5,197 +5,187 @@ class SongListener {
         this.playlists = this.app.playlists;
         this.mainFolder = this.app.mainFolder;
 
-        this.elements = {
-            audio: document.getElementById("audio"),
-        };
-        this.elements.audio.volume = 0;
-        this.queue = null;
+        this.audio = document.getElementById("audio");
+        this.audio.volume = 0;
+
+        this.queue = [];
+        this.addedToQueue = [];
         this.currentPlaylist = null;
-        this.lastError = "";
-
+    
         this.handleEvents();
-    }
-
-    error(message) {
-        this.app.error(message);
-        this.lastError = message;
-        setTimeout(() => this.lastError = "", 10);
+        this.initQueues();
     }
 
     handleEvents() {
-        this.elements.audio.addEventListener("ended", () => this.next());
-        this.elements.audio.addEventListener("error", (e) => {
-            if (e.target.error) this.playNextQueueSong();
+        this.audio.addEventListener("ended", () => this.playNextSong(1));
+        this.audio.addEventListener("error", (e) => {
+            if (e.target.error) this.app.error("ERROR HK-303 =>", e.target.error);
         });
     }
-    
-    resetQueue() {
-        this.queue = this.currentPlaylist.songs.map((sID) => {
-            return {
-                id: sID,
-                manuallyAdded: false,
-            };
-        });
+
+    initQueues() {
+        const params = new URLSearchParams(window.location.search);
+        const jsonData = params.get("d");
+
+        if (jsonData) {
+            const d = JSON.parse(decodeURIComponent(jsonData));
+            if (!d.playlist) return;
+
+            this.currentPlaylist = d.playlist;
+            this.queue = d.queue;
+            this.addedToQueue = d.addedToQueue;
+            
+            this.queue.unshift(d.songID);
+            this.playNextSong(1);
+            this.setSongCurrentTime(d.currentTime);
+        }
+    }
+
+    resetQueue(startIndex) {
+        const random = this.app.settings.random;
+        const loop = this.app.settings.loop;
+
+        this.queue = [];
+        if (this.currentPlaylist.songs.length == 0) return;
+
+        if (loop) this.queue = [this.currentPlaylist.songs[startIndex]];
+
+        else if (random) {
+            this.queue = this.currentPlaylist.songs.map(() => {
+                return this.currentPlaylist.songs[rand(0, this.currentPlaylist.songs.length - 1)];
+            });
+
+        } else {
+            this.queue = [...this.currentPlaylist.songs];
+            this.shiftQueue(startIndex);
+        }
     }
 
     shiftQueue(incr) {
         for (let i = 0; i < Math.abs(incr); i++) {
-            if (incr > 0) this.queue.push(this.queue.shift());
-            else this.queue.unshift(this.queue.pop());
+            if (!this.app.settings.random) {
+                if (incr > 0) this.queue.push(this.queue.shift());
+                else if (incr < 0) this.queue.unshift(this.queue.pop());
+            } else {
+                const randSongID = this.currentPlaylist.songs[rand(0, this.currentPlaylist.songs.length - 1)];
+                if (incr > 0) {
+                    this.queue.shift();
+                    this.queue.push(randSongID);
+                } else if (incr < 0) {
+                    this.queue.pop();
+                    this.queue.unshift(randSongID);
+                }
+            }
         }
     }
 
-    updateQueue(incr) {
-        if (!this.app.settings.loop) {
-            if (!this.app.settings.random) {
-                if (this.queue) this.shiftQueue(incr);
-                else this.resetQueue();
-
-            } else {
-                const randomID = this.currentPlaylist.songs[rand(0, this.currentPlaylist.songs.length - 1)];
-
-                if (incr > 0) {
-                    this.queue.shift();
-                    this.queue.push({
-                        id: randomID,
-                        manuallyAdded: false,
-                    });
-                } else {
-                    this.queue.pop();
-                    this.queue.unshift({
-                        id: randomID,
-                        manuallyAdded: false,
-                    });
-                }
-            }
-        } else this.queue = this.currentPlaylist.songs.map(() => {
-            return {
-                id: (this.app.currentSondID == -1) ? 0 : this.app.currentSondID,
-                manuallyAdded: false,
-            };
-        });
+    setCurrentPlaylist(playlist, startIndex, incr) {
+        this.currentPlaylist = playlist;
+        this.resetQueue(startIndex);
+        this.playNextSong(incr);
     }
 
-    play() {
-        const audio = this.elements.audio;
-
-        if (this.app.currentSondID != -1) {
-            if (this.isPaused()) {
-                audio.play();
-                this.app.elements.footer.buttons.pause.classList.remove("hidden");
-                this.app.elements.footer.buttons.play.classList.add("hidden");
-            } else {
-                audio.pause();
-                this.app.elements.footer.buttons.play.classList.remove("hidden");
-                this.app.elements.footer.buttons.pause.classList.add("hidden");
-            }
-        } else this.playNextQueueSong();
-    }
-
-    previous() {
+    playNextSong(incr) {
         if (this.queue.length == 0) return;
-        const audio = this.elements.audio;
+        if (incr > 1) incr = 1;
+        else if (incr < -1) incr = -1;
+        
+        const sID = (this.addedToQueue.length == 0) ? this.queue.at((incr > 0) ? 0 : -1) : this.addedToQueue[0];
 
-        if (audio.currentTime < 5) this.playPreviousQueueSong();
-        else audio.currentTime = 0;
-    }
+        if (this.addedToQueue.length != 0) this.addedToQueue.shift();
+        else this.shiftQueue(incr);
 
-    next() {
-        this.playNextQueueSong();
-    }
+        const song = this.songs[sID];
+        this.audio.src = path.join(this.mainFolder, "songs", song.src);
+        this.audio.setAttribute("song-id", sID);
+        this.audio.play();
 
-    random() {
-        this.app.settings.random = !this.app.settings.random;
-        if (this.app.settings.random) this.queue = this.currentPlaylist.songs.map(() => {
-            return {
-                id: this.currentPlaylist.songs[rand(0, this.currentPlaylist.songs.length - 1)],
-                manuallyAdded: false,
-            };
-        }); else this.resetQueue();
-    }
-
-    loop() {
-        this.app.settings.loop = !this.app.settings.loop;
-        if (!this.app.settings.loop) {
-            this.resetQueue();
-            if (this.app.settings.random) {
-                this.app.settings.random = !this.app.settings.random;
-                this.random();
-            }
-            this.shiftQueue(this.currentPlaylist.songs.indexOf(this.app.currentSondID) + 1);
-        } else this.updateQueue(0);
+        this.app.elements.footer.buttons.pause.classList.remove("hidden");
+        this.app.elements.footer.buttons.play.classList.add("hidden");
     }
 
     addToQueue(sID) {
-        this.queue.unshift({ id: sID, manuallyAdded: true });
+        this.addedToQueue.push(parseInt(sID));
     }
 
-    playNextQueueSong() {
-        if (this.queue.length == 0) return;
-        this.playSong(this.queue[0].id);
 
-        if (this.queue[0].manuallyAdded) this.queue.shift();
-        else this.updateQueue(1);
+
+    randomButton() {
+        if (!this.currentPlaylist) return;
+        const currentSid = this.getCurrentSongID();
+        const currentSidIndex = this.currentPlaylist.songs.indexOf(currentSid);
+        this.resetQueue(currentSidIndex + 1);
     }
 
-    playPreviousQueueSong() {
-        this.updateQueue(-1);
-        this.playSong(this.queue.at(-1).id);
+    loopButton() {
+        if (!this.currentPlaylist) return;
+        const currentSid = this.getCurrentSongID();
+        const currentSidIndex = this.currentPlaylist.songs.indexOf(currentSid) + ((this.app.settings.loop) ? 0 : 1);
+        this.resetQueue(currentSidIndex);
     }
 
-    playSong(id) {
-        const song = this.songs[id];
-
-        if (song) {
-            if (fs.existsSync(path.join(this.mainFolder, "songs", song.src))) {
-                const audio = this.elements.audio;
-                audio.src = `${this.mainFolder}/songs/${song.src}`;
-                audio.play();
-                this.app.elements.footer.buttons.pause.classList.remove("hidden");
-                this.app.elements.footer.buttons.play.classList.add("hidden");
-                this.app.currentSondID = id;
-            } else this.error(`File "${song.src}" is missing for the song "${song.name} by ${song.artist}"`);
-        } else this.error(`Song ID "${id}" not found`);
+    playButton() {
+        if (this.isSrcValid()) {
+            if (this.isPaused()) this.audio.play();
+            else this.audio.pause();    
+        } else this.setCurrentPlaylist(this.app.currentPlaylist, 0, 1);
     }
+
+    nextButton() {
+        if (this.currentPlaylist) this.playNextSong(1);
+        else this.setCurrentPlaylist(this.app.currentPlaylist, 0, 1);
+    }
+
+    previousButton() {
+        if (this.currentPlaylist) {
+            if (this.getCurrentSongCurrentTime() > 5) this.setSongCurrentTime(0);
+            else this.playNextSong(-1);
+        }
+        else this.setCurrentPlaylist(this.app.currentPlaylist, 0, -1);
+    }
+
+
 
     isPaused() {
-        return audio.paused;
+        return this.audio.paused;
     }
 
-    getError() {
-        return this.lastError;
+    isSrcValid() {
+        return this.audio.src != "";
     }
+
+
 
     getCurrentSongDuration() {
-        return this.elements.audio.duration;
+        return this.audio.duration;
     }
 
     getCurrentSongCurrentTime() {
-        return this.elements.audio.currentTime;
+        return this.audio.currentTime;
     }
 
-    getCurrentPlaylist() {
-        return this.currentPlaylist;
+    getCurrentSongID() {
+        const sID = parseInt(this.audio.getAttribute("song-id"));
+        return isNaN(sID) ? -1 : sID;
+    }
+
+    getCurrentData() {
+        return {
+            playlist: this.currentPlaylist,
+            songID: this.getCurrentSongID(),
+            currentTime: this.getCurrentSongCurrentTime(),
+            queue: this.queue,
+            addedToQueue: this.addedToQueue,
+        };
+    }
+
+
+
+    setSongCurrentTime(time) {
+        this.audio.currentTime = time;
     }
 
     setVolume(volume) {
-        this.elements.audio.volume = volume;
-    }
-
-    setCurrentTime(time) {
-        this.elements.audio.currentTime = time;
-    }
-
-    setCurrentPlaylist(playlist) {
-        this.currentPlaylist = playlist;
-        this.resetQueue();
-        this.updateQueue(0);
-        if (this.app.settings.random) {
-            this.app.settings.random = !this.app.settings.random;
-            this.random();
-        } else if (this.app.settings.loop) {
-            this.app.settings.random = !this.app.settings.random;
-            this.loop();
-        }
+        this.audio.volume = volume;
     }
 };